Adjusted strawberry and blueberry bush spawn locations

+++ src/fruitwars/Game.java
@@ -61,13 +61,13 @@ public class Game extends Scene
 		StrawberryBush s = new StrawberryBush(entities, new Point2D.Double(100, 100), true, 150);
 		s.setRally(new Point2D.Double(600, 300));
 		entities.add(s);
-		s = new StrawberryBush(entities, new Point2D.Double(600, 500), false, 150);
+		s = new StrawberryBush(entities, new Point2D.Double(600, 100), false, 150);
 		s.setRally(new Point2D.Double(600, 300));
 		entities.add(s);
 		BlueberryBush b = new BlueberryBush(entities, new Point2D.Double(100, 500), true, 150);
 		b.setRally(new Point2D.Double(100, 300));
 		entities.add(b);
-		b = new BlueberryBush(entities, new Point2D.Double(600, 100), false, 150);
+		b = new BlueberryBush(entities, new Point2D.Double(600, 500), false, 150);
 		b.setRally(new Point2D.Double(100, 300));
 		entities.add(b);
 		GrapeVine g = new GrapeVine(entities, new Point2D.Double(100, 300), true, 150);
